move plugins dir one level up
