+++ src/core/flow.js
@@ -12,7 +12,7 @@ import {
   UpdateTurnOrderState,
   TurnOrder,
 } from './turn-order';
-import { FnWrap } from './plugins/main';
+import { FnWrap } from '../plugins/main';
 import { automaticGameEvent } from './action-creators';
 import { ContextEnhancer } from './reducer';
 import * as logging from './logger';
