+++ src/plugins/plugin-player.js
